auth/views.py
from django.contrib.auth.models import User, Group
from .serializers import UserSerializer, GroupSerializer
from rest_framework import permissions, viewsets
from rest_framework_simplejwt.authentication import JWTStatelessUserAuthentication, JWTAuthentication


class UserViewSet(viewsets.ModelViewSet):
    """
    API endpoint that allows users to be viewed or edited.
    """

    queryset = User.objects.all().order_by("-date_joined")
    serializer_class = UserSerializer
    permission_classes = [permissions.IsAuthenticated]
    authentication_classes = [JWTStatelessUserAuthentication]


class GroupViewSet(viewsets.ModelViewSet):
    """
    API endpoint that allows groups to be viewed or edited.
    """

    queryset = Group.objects.all().order_by("name")
    authentication_classes = [JWTStatelessUserAuthentication]
    serializer_class = GroupSerializer
    permission_classes = [permissions.IsAuthenticated]
